Expanding petsc with matvec_copy_{col,row}. Add ACTIVE test for plugins.
OPAQUE_ACTIVE is set to false for all plugin objects.
If the initializer function sets it to TRUE, then
the object can not be removed from memory via a restore until
it's reset manually or by KILL_OPAQUE.

This is checked in check_plugins in restore which is active for
dnodes && dpawns (it's a function pointer to _check_plugins).

matvec_copy_* funcs are for icc.