Cleaning up the file descriptor fixes.
a stdin member is added to the SocketType
structures to keep stdin, stdout from being
merely closed, but instead they are redirected
from /dev/null. Then dqueen dups stdout to stderr
so that mpi output has an outlet.

Also, sigfd close-on-exec flags was fixed, and
a few strings pointing to old ".*.in" files were
updated to ".*in"